+++ __tests__/services/grpc/world_engine_client_grpc.service.spec.ts
@@ -9,7 +9,6 @@ import {
   ClientRequest,
   LoadSceneResponse,
 } from '@proto/world-engine_pb';
-import { v4 } from 'uuid';
 
 import { Config } from '../../../src/common/config';
 import { CLIENT_ID } from '../../../src/common/constants';
@@ -17,11 +16,14 @@ import { WorldEngineClientGrpcService } from '../../../src/services/gprc/world_e
 import {
   createAgent,
   extension,
+  KEY,
+  SCENE,
+  SECRET,
   sessionContinuation,
+  sessionProto,
   sessionToken,
   user,
 } from '../../helpers';
-const SCENE = v4();
 
 const agents = [createAgent(), createAgent()];
 
@@ -35,7 +37,7 @@ describe('credentials', () => {
 
   test('should use insecure credentials', () => {
     const generateSessionToken = jest
-      .spyOn(Config.prototype, 'getEngineSsl')
+      .spyOn(Config.prototype, 'getSsl')
       .mockImplementationOnce(() => false);
 
     new WorldEngineClientGrpcService();
@@ -47,7 +49,7 @@ describe('credentials', () => {
 
   test('should use secure credentials', () => {
     const generateSessionToken = jest
-      .spyOn(Config.prototype, 'getEngineSsl')
+      .spyOn(Config.prototype, 'getSsl')
       .mockImplementationOnce(() => true);
 
     new WorldEngineClientGrpcService();
@@ -55,6 +57,30 @@ describe('credentials', () => {
     expect(createSsl).toHaveBeenCalledTimes(1);
     expect(createInsecure).toHaveBeenCalledTimes(0);
     expect(generateSessionToken).toHaveBeenCalledTimes(1);
+  });
+});
+
+describe('generateSessionToken', () => {
+  test('should generate token', async () => {
+    const generateSessionToken = jest
+      .spyOn(WorldEngineClient.prototype, 'generateToken')
+      .mockImplementationOnce((_request, _metadata, _options, callback) => {
+        const cb = typeof _options === 'function' ? _options : callback;
+        cb(null, sessionProto);
+        return {} as SurfaceCall;
+      });
+
+    const service = new WorldEngineClientGrpcService();
+    const result = await service.generateSessionToken(
+      {
+        key: KEY,
+        secret: SECRET,
+      },
+      SCENE,
+    );
+
+    expect(generateSessionToken).toHaveBeenCalledTimes(1);
+    expect(result).toEqual(sessionProto);
   });
 });