+++ src/clients/inworld.client.ts
@@ -6,6 +6,7 @@ import {
   UserRequest,
 } from '@proto/world-engine_pb';
 
+import { SCENE_PATTERN } from '../common/constants';
 import {
   ApiKey,
   Awaitable,
@@ -102,10 +103,11 @@ export class InworldClient<
   }
 
   async generateSessionToken() {
-    this.validateApiKey();
+    this.validate();
 
     return new ConnectionService({
       apiKey: this.apiKey,
+      name: this.scene,
     }).generateSessionToken();
   }
 
@@ -167,14 +169,22 @@ export class InworldClient<
     }
   }
 
+  private validateScene() {
+    if (!this.scene) {
+      throw Error('Scene name is required');
+    }
+
+    if (!SCENE_PATTERN.test(this.scene)) {
+      throw Error('Scene name has wrong format');
+    }
+  }
+
   private validate() {
     if (!this.generateSessionTokenFn) {
       this.validateApiKey();
     }
 
-    if (!this.scene) {
-      throw Error('Scene name is required');
-    }
+    this.validateScene();
   }
 }